[feat] 테스트 코드 파일 위치 이동
